Reorganize phone utils
